Add files via upload
The knowledge of historical booking data can assist airlines to determine and anticipate future reservation from the previously available data. The advance booking model uses the airline reservations to predict the demand on a given day and looks at the cumulative bookings for a particular future flight as they come in. The basis of this approach is a booking curve, a concept which provides the cumulative advance bookings made against departure date.  

Using the booking curve, data points such as, days prior to departure date, day of week and demand can be derived. These data points are then directly used for forecasting the final reservation demand using two models.